Corrected type for scale in Effects. Pass default param to Sprite. Added more details to Cell and Grid to load and render.

+++ src/mugen/mugen_util.h
@@ -87,8 +87,8 @@ class Effects{
 	bool mask;
 	int facing;
 	int vfacing;
-	int scalex;
-	int scaley;
+	double scalex;
+	double scaley;
 };
 
 /* Use with fonts */
